Refactor command methods in more manageable modules
