+++ bin/cluster_creator
@@ -2,6 +2,7 @@
 
 # frozen_string_literal: true
 
+$LOAD_PATH.unshift(File.expand_path('../lib', __dir__))
 require_relative '../test/support/cluster/orchestrator'
 
 urls = ARGV.map { |host_port| "redis://#{host_port}" }
